fixup! enable newer git in github actions by actions/checkout

+++ dockerfiles/util/bootstrap.inc.sh
@@ -24,4 +24,4 @@ make VERSION
 git rev-parse HEAD > /support-firecloud.bootstrapped
 
 # enable newer git in github actions by actions/checkout, or else no history/submodules/etc with REST API checkouts
-ln -s /home/linuxbrew/.linuxbrew/bin/git /usr/local/bin/git
+[[ -e /usr/local/bin/git ]] || ln -s /home/linuxbrew/.linuxbrew/bin/git /usr/local/bin/git
